Login & Out Functionality Working But..
While it is working, for Brave (when shields are up) and a bunch of
protections are in place, the Google Sign In does not seem to work.
As a Temporary Measure I added an alert to request the user to allow
for 3rd Party Cookies etc. [Or Just Turning off the Shield] when
accessing this site. We can look into this issue in the future.