Individual Note Basic Editing (#10)
* Rudimentary UI Done

* Working Markdown Renderer with Placeholder Text

* Linked the TextArea to the MarkdownRenderer and Created Note Model

The gist of the rendering and translation works. Now to link the
individual note to the redux store.

* Working Markdown Interpretation + Redux Compatibility

The app can take a note, edit it with a realtime rendering of edited
content, and save it back to the store. For now this is a rudimentary
working version of the "Note Editing" feature.